[Refactor] LocationService 프로토콜 생성, 구현체 분리
